pyproject.toml
[project]
name = "grpcdemo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
	"grpcio",
	"grpcio-tools",
	"protobuf",
]
